Generator

+++ token_types.h
@@ -114,8 +114,8 @@ bool isTerminal(tokenType type);
 typedef union Value_of_token
 {
     int INT_VAL;
-    float FLOAT_VAL;
-    char *STR_VAL;
+    double FLOAT_VAL;
+    char* STR_VAL;
 } value_of_token;
 
 // structure of token
